pagination and file manipulations

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require("../models/product");
 const util = require("util");
+const fileHelper = require("../util/file");
 const mongodb = require("mongodb");
 const {
   validationResult
@@ -24,12 +25,29 @@ exports.getAddProduct = (req, res, next) => {
 exports.addProduct = async (req, res, next) => {
 
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  // const imageUrl = req.body.imageUrl;
+  const image = req.file
   const price = req.body.price;
   const description = req.body.description;
+  if(!image) {
+    return res.status(422).render("admin/edit-product", {
+      docTitle: "Add Product",
+      btnTitle: "Add Product",
+      labelTitle: "Title",
+      path: "/admin/add-product",
+      editing: false,
+      hasError: true,
+      product: {
+        title,
+        price,
+        description
+      },
+      errorMsg: 'Attached file is not an image'
 
+
+    });
+  }
   const errors = validationResult(req);
-  console.log(errors)
   if (!errors.isEmpty())
     return res.status(422).render("admin/edit-product", {
       docTitle: "Add Product",
@@ -40,7 +58,6 @@ exports.addProduct = async (req, res, next) => {
       hasError: true,
       product: {
         title,
-        imageUrl,
         price,
         description
       },
@@ -50,6 +67,7 @@ exports.addProduct = async (req, res, next) => {
     });
 
   try {
+    const imageUrl = image.path;
     const product = new Product({
       title,
       imageUrl,
@@ -63,6 +81,9 @@ exports.addProduct = async (req, res, next) => {
     console.log('Product has been saved', savedProd);
   } catch (error) {
     console.log("Save operation failed ", error.message);
+    const err = new Error(error.message);
+    err.httpStatus = 500;
+    return next(err)
   }
 };
 
@@ -75,6 +96,7 @@ exports.getEditProduct = async (req, res, next) => {
   const prodId = req.params.productId;
   try {
     product = await Product.findById(prodId);
+    console.log(product)
     res.render("admin/edit-product", {
       docTitle: "Edit Product",
       btnTitle: "Update Product",
@@ -87,6 +109,7 @@ exports.getEditProduct = async (req, res, next) => {
     });
   } catch (error) {
     console.log("Edit product failed ", error.message);
+    return next(err)
   }
 };
 
@@ -94,9 +117,8 @@ exports.postEditProduct = async (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
-
   const errors = validationResult(req);
 
   if (!errors.isEmpty()){
@@ -109,9 +131,9 @@ exports.postEditProduct = async (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
-        description: updatedDesc
+        description: updatedDesc,
+        _id: prodId,
       },
       errorMsg: errors.array()[0].msg
     });
@@ -124,8 +146,12 @@ exports.postEditProduct = async (req, res, next) => {
     }
     product.title = updatedTitle;
     product.price = updatedPrice;
-    product.imageUrl = updatedImageUrl;
     product.description = updatedDesc;
+    if(image){
+      fileHelper.deleteFile(product.imageUrl);
+      
+      product.imageUrl = image.path
+    }
     const object_id = await product.save();
     console.log("Product updated", object_id);
     res.redirect("/admin/products");
@@ -151,12 +177,19 @@ exports.getProducts = async (req, res, next) => {
 };
 
 exports.deleteProduct = async (req, res, next) => {
-  const prodId = req.body.productId;
+
   try {
+    const prodId = req.body.productId;
+    const prod = await Product.findById(prodId);
+    if(!prod){
+      return next(new Error('Product not found'));
+    }
     const product = await Product.deleteOne({
       _id: prodId,
       userId: req.user._id
     })
+    fileHelper.deleteFile(prod.imageUrl);
+    
     console.log("Product deleted", product);
     res.redirect("/admin/products");
   } catch (error) {
